don't quote me boy

+++ MissTortasRework/MissTortas.Domain/Orders/OrderSaleProduct.cs
@@ -6,7 +6,7 @@ namespace MissTortas.Domain.Orders
     {
         public long OrderSaleProductId { get; set; }
         public long OrderId { get; set; }
-        public double QuantityAsked { get; set; }
+        public decimal QuantityAsked { get; set; }
         public long SaleProductId { get; set; }
         public virtual SaleProduct SaleProduct { get; set; } = null!;
         public virtual Order Order { get; set; } = null!;
