+++ MissTortasRework/MissTortas.Domain/Products/Discount.cs
@@ -3,7 +3,7 @@ namespace MissTortas.Domain.Products
     public class Discount
     {
         public int Percentage { get; set; }
-        public float Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public string Reason { get; set; } = string.Empty;
         public required SaleProduct SaleProduct { get; set; }
     }
